+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+- Fixed archetype refresh bugs and changed semantics of `refresh_archetype`.
 - Fixed potential "cannot shadow tuple struct with let binding" errors in derived `Query` and `SystemParam` impls.
+- Added `#[must_use]` attribute to iterators.
+- Relaxed auto trait impls for various types.
 - Minor performance improvements.
 
 ## 0.1.1 - 2024-01-25